Refactor number manipulation into a lib module
