+++ packages/components/src/index.ts
@@ -39,10 +39,7 @@ export {
 	StatsCardAvatar,
 	StatsCardTitleExtras,
 } from './horizontal-bar-list';
-export {
-	default as CountComparisonCard,
-	percentCalculator,
-} from './highlight-cards/count-comparison-card';
+export { default as CountComparisonCard } from './highlight-cards/count-comparison-card';
 export { default as AnnualHighlightCards } from './highlight-cards/annual-highlight-cards';
 export {
 	default as WeeklyHighlightCards,
